+++ lib/pludoni_rspec/spec_helper.rb
@@ -1,4 +1,7 @@
 RSpec.configure do |config|
+  config.before(:each) do
+    ActiveSupport::CurrentAttributes.reset_all if defined?(ActiveSupport::CurrentAttributes)
+  end
   config.example_status_persistence_file_path = 'tmp/rspec.failed.txt'
   config.expect_with :rspec do |expectations|
     expectations.include_chain_clauses_in_custom_matcher_descriptions = true
@@ -62,4 +65,37 @@ RSpec.configure do |config|
       example.run
     end
   end
+
+  # helper method for below
+  def current_file_is_the_only_spec?
+    file = caller.find { |c| c =~ /_spec\.rb/ }.split(':').first
+    file = Pathname.new(file).relative_path_from(Rails.root).to_s
+    ARGV.any? { |arg| arg.include?(file) }
+  end
+
+  # skip this example, only run if you are working on it and skip it later
+  #   like special customers bugs etc.
+  def only_run_when_single_spec_and_local!
+    if ENV['CI'] || !current_file_is_the_only_spec?
+      skip "Only run when run as a single spec and not CI set"
+    end
+  end
+
+  # For System specs:
+  # prints the instructions to open the current page in your local browser
+  # if you run the specs remotely, create a SSH tunnel to the server during system specs
+  # and open the brOwser on your local machine
+  def local!
+    uri = URI.parse(page.current_url)
+    puts "--- Connect with local browser:"
+    puts "  1. Open a SSH tunnel with port forwarding to the test server:"
+    puts "\nssh #{ENV['LOGNAME']}@pludoni.com -L 8080:localhost:#{uri.port}\n"
+      puts "  2. Open in Browser: "
+    uri.port = nil
+    uri.scheme = nil
+    uri.host = nil
+    puts "\nhttp://localhost:8080#{uri}\n"
+    puts "  Afterwards, you can close the SSH session"
+  end
+
 end